Task11/user boilerplate (#12)
* update user schema to contain city, state, and isAcceptingRequests

* update chapter and user routes

* update user schema with isActive field

* fix register function for admin to add new chapter

* fixing register function
