+++ server/src/controllers/birthdayRequest.controller.ts
@@ -11,7 +11,7 @@ import {
   deleteRequestByID,
   createBirthdayRequestByID,
 } from '../services/birthdayRequest.service.ts';
-import { getChapterById } from '../services/chapter.service.ts';
+import { getUserById } from '../services/user.service.ts';
 import {
   emailRequestUpdate,
   emailRequestDelete,
@@ -22,7 +22,7 @@ import {
   ChildSituation,
   IBirthdayRequest,
 } from '../models/birthdayRequest.model.ts';
-import { IChapter } from '../models/chapter.model.ts';
+import { IUser } from '../models/user.model.ts';
 
 const getAllRequests = async (
   req: express.Request,
@@ -74,7 +74,7 @@ const updateRequestStatus = async (
     return;
   }
   // get chapter email by chapter ID
-  const chapter: IChapter | null = await getChapterById(request.chapterId);
+  const chapter: IUser | null = await getUserById(request.chapterId);
   if (!chapter) {
     next(ApiError.notFound(`Chapter does not exist`));
     return;
@@ -128,7 +128,7 @@ const deleteRequest = async (
     return;
   }
   // get chapter email by chapter ID
-  const chapter: IChapter | null = await getChapterById(request.chapterId);
+  const chapter: IUser | null = await getUserById(request.chapterId);
   if (!chapter) {
     next(ApiError.notFound(`Chapter does not exist`));
     return;
