+++ server/src/routes/user.route.ts
@@ -1,23 +1,12 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import express from 'express';
 import { isAdmin } from '../controllers/admin.middleware.ts';
-import {
-  toggleRequest,
-  getAllChapters,
-  createChapter,
-  deleteChapter,
-} from '../controllers/chapter.controller.ts';
+import { toggleRequest } from '../controllers/user.controller.ts';
 import { isAuthenticated } from '../controllers/auth.middleware.ts';
 import 'dotenv/config';
 
 const router = express.Router();
 
 router.put('/toggleRequests/:id', isAuthenticated, isAdmin, toggleRequest);
-
-router.get('/all', isAuthenticated, isAdmin, getAllChapters);
-
-router.post('/create/', isAuthenticated, isAdmin, createChapter);
-
-router.delete('/delete/:id', isAuthenticated, isAdmin, deleteChapter);
 
 export default router;
